+++ sdk/python/pulumi_keycloak/group_memberships.py
@@ -134,7 +134,7 @@ class GroupMemberships(pulumi.CustomResource):
         """
         ## Import
 
-        This resource does not support import. Instead of importing, feel free to create this resource as if it did not already exist on the server.
+        This resource does not support import. Instead of importing, feel free to create this resource as if it did not already exist on the server. [1]/docs/resources/user_groups.html
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
@@ -151,7 +151,7 @@ class GroupMemberships(pulumi.CustomResource):
         """
         ## Import
 
-        This resource does not support import. Instead of importing, feel free to create this resource as if it did not already exist on the server.
+        This resource does not support import. Instead of importing, feel free to create this resource as if it did not already exist on the server. [1]/docs/resources/user_groups.html
 
         :param str resource_name: The name of the resource.
         :param GroupMembershipsArgs args: The arguments to use to populate this resource's properties.